Fix Wrong parse int
Fix Wrong parse int

+++ src/components/modal/input.jsx
@@ -32,7 +32,7 @@ export default function CompraInput({
             store,
             date: date + '/' + month.date,
             categoria,
-            price: parseInt(price)
+            price: parseFloat(price)
         };
         const newMonth = await addCompra(compra, month, index);
         setMonth(newMonth);
